pyproject.toml
[tool.poetry]
name = "votepizza"
version = "0.1.0"
description = "RESTful app for voting for the best pizza"
authors = ["Alex Rozum <rozumalex@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.2"
flake8 = "^3.8.4"
djangorestframework = "^3.12.1"
django-cors-headers = "^3.5.0"
django-environ = "^0.4.5"
psycopg2 = "^2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
